Make it compatible with python==3.10 and cuda==12.1

+++ loops/sampler.py
@@ -65,7 +65,7 @@ class MultiClassSampler:
         for each line we search in the index for the correct label and assign 1 in the resulting vector
         """
         # statement shape for correct processing of the very last batch which size might be less than self.bs
-        y = np.zeros((statements.shape[0], self.n_entities), dtype=np.float32)
+        y = np.zeros((statements.shape[0], self.n_entities), dtype=float)
 
 
         for i, s in enumerate(statements):
